# demo/sever/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NodeJS API for IOT Dashboard",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:2002/"
      }
    ],
    "paths": {
      "/data_sensor/all": {
        "get": {
          "summary": "Gọi tất cả các Data Sensor",
          "description": "Gọi tất cả các Data Sensor bao gồm việc phân trang và sắp xếp",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number",
              "required": false,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "sort",
              "in": "query",
              "description": "Field to sort by ('temperature', 'humidity', 'light', 'created_at') or 'all' (default: 'id')",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["temperature", "humidity", "light", "created_at", "all"]
              }
            },
            {
              "name": "order",
              "in": "query",
              "description": "Sort order ('ASC' or 'DESC')",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["ASC", "DESC"]
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "The number of items per page",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": "1",
                "default": "10"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "404": {
              "description": "No data found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/data_sensor/search": {
        "get": {
          "summary": "Tìm kiếm dữ liệu trong Data Sensor",
          "description": "Tìm kiếm dữ liệu trong Data Sensor bao gồm việc phân trang và lọc",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number",
              "required": false,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "filter",
              "in": "query",
              "description": "Field to filter by (all, temperature, humidity, light, created_at)",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["all", "temperature", "humidity", "light", "created_at"]
              }
            },
            {
              "name": "value",
              "in": "query",
              "description": "Value to filter by",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "The number of items per page",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": "1",
                "default": "10"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "To test Get method"
            }
          }
        }
      },
      
      "/data_sensor/addData": {
        "post": {
          "summary": "Thêm Data Sensor",
          "description": "Thêm Data Sensor mới vào trong database",
          "tags": ["Data Sensor"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "light": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Data added successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/data_sensor/{id}": {
        "put": {
          "summary": "Cập nhật Data Sensor",
          "description": "Cập nhật Data Sensor vào database qua ID",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "ID của Data Sensor cần cập nhật",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "light": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Data updated successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "404": {
              "description": "Data not found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "summary": "Xóa Data Sensor",
          "description": "Xóa Data Sensor vào database qua ID",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "ID của Data Sensor cần xóa",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Data deleted successfully"
            },
            "404": {
              "description": "Data not found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }     ,

      "/action_history/all": {
        "get": {
          "summary": "Gọi tất cả các Action History",
          "description": "Gọi tất cả các Action History bao gồm việc phân trang và sắp xếp",
          "tags": ["Action History"],
          "parameters": [
            {
              "name": "page",
              "in": "path",
              "description": "Page number",
              "required": true,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "sort",
              "in": "query",
              "description": "Sort by field ('Device', 'Action', 'created_at') or 'all' (default: 'id')",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["Device", "Action", "created_at", "all"]
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "The number of items per page",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": "1",
                "default": "10"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "404": {
              "description": "No data found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },

      "/action_history/search": {
        "get": {
          "summary": "Tìm kiếm dữ liệu trong Action History",
          "description": "Tìm kiếm dữ liệu trong Action History bao gồm việc phân trang và lọ",
          "tags": ["Action History"],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number",
              "required": false,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "filter",
              "in": "query",
              "description": "Field to filter by ('all', 'device', 'action')",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["all", "device", "action"]
              }
            },
            {
              "name": "value",
              "in": "query",
              "description": "Value to filter by",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "The number of items per page",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": "1",
                "default": "10"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "404": {
              "description": "No data found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/action_history/addData": {
        "post": {
          "summary": "Thêm Action History",
          "description": "Thêm Action History mới vào database",
          "tags": ["Action History"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Device": {
                      "type": "string"
                    },
                    "action": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Data added successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/action_history/{id}": {
        "put": {
            "summary": "Cập nhật Action History",
            "description": "Cập nhật Action History vào database qua ID",
            "tags": ["Action History"],
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "ID của Action History cần cập nhật",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "device": {
                                    "type": "string"
                                },
                                "action": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Data updated successfully"
                },
                "400": {
                    "description": "Bad request. Missing required fields or invalid data"
                },
                "404": {
                    "description": "Data not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            }
        },
        "delete": {
          "summary": "Xóa Action History",
          "description": "Xóa Action History vào database qua ID",
          "tags": ["Action History"],
          "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID của Action History cần xóa",
                  "required": true,
                  "schema": {
                      "type": "integer"
                  }
              }
          ],
          "responses": {
              "200": {
                  "description": "Data deleted successfully"
              },
              "404": {
                  "description": "Data not found"
              },
              "500": {
                  "description": "Internal server error"
              }
          }
      }
      
    }
    
  
    
      
    }      

}